[update] コメント検索・削除ロジックをユーティリティ関数として分離
コメント検索と削除処理を汎用化し、`commentUtils.js` に移動しました。これによりコードの再利用性と可読性が向上し、Vue コンポーネント側でのロジックの簡潔化を実現しました。
